incluiir commponentização
