+++ src/components/SettingsModal.jsx
@@ -1,0 +1,20 @@
+import React from 'react'
+
+export default function SettingsModal({ onClose }) {
+  return (
+    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
+      <div className="bg-white dark:bg-gray-800 p-6 rounded shadow-lg w-96">
+        <h2 className="text-lg font-bold mb-4">Configurações</h2>
+        <p className="text-sm text-gray-600 dark:text-gray-300">Em breve você poderá configurar opções aqui.</p>
+        <div className="mt-4 flex justify-end">
+          <button
+            onClick={onClose}
+            className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600"
+          >
+            Fechar
+          </button>
+        </div>
+      </div>
+    </div>
+  )
+}
